Hardy Cross Method implementation
Began implementing the Hardy Cross Method. Loops, logic checks, and calculations are established.

In progress:
Need to add the implementation of flows in and out into the current Hardy Cross Method loops.
Need to implement writing the final data to a text file.
A few warnings on improperly defined functions need to be cleaned up.
In the future, loops could be procedurally detected instead of requiring them to be specified. (See loopDefine function.)