More work on Hardy Cross method implementation
More work on Hardy Cross method implementation. Currently converging, however, it does not currently solve the problem.

To do: Implement flow in and out of the system. This is the final element of the implementation of the Hardy Cross method. After implementing flow in and out, the problem will be solved.