dQ and Q calculations, refactoring
Refactored the project slightly by modifying a few class structures to potentially make it easier to implement the input and output flows in the future.

Moved all the code from the HardyCross class into the main method to make the necessary global variables work easier.

To do: Ensure that hL converges and is being recorded properly (hL calculation is not currently working correctly). Implement input and output flows (mass balance).